Changed all ints to Bytes
Cleaned up c files to be more efficient. Added Comments galore.

+++ Arduino_Library/SerialGraphicLCD/SerialGraphicLCD.h
@@ -30,13 +30,16 @@ class LCD
 	void setHome();
 	void demo();
 	void setPixel(int x, int y, int set);
-	void drawLine(int x1, int y1, int x2, int y2, int set);
+	void drawLine(byte x1, byte y1, byte x2, byte y2, byte set);
 	void drawBox(int x1, int y1, int x2, int y2, int set);
 	void drawCircle(int x, int y, int rad, int set);
 	void eraseBlock(int x1, int y1, int x2, int y2);
 	
 	
 	private:
+    
+    
+    
 
 };
